backend/pyproject.toml
[project]
name = "scalecraft-backend"
version = "0.1.0"
description = "ScaleCraft FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "networkx>=3.3,<4.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
]
openai = [
  "openai>=1.52.0,<2.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
